requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
djoser
social-auth-app-django==4.0.0